user functionality changes

+++ server/controllers/posts.js
@@ -3,7 +3,7 @@ import PostMessage from "../models/postMessage.js";
 export const getPosts = async (req, res) => {
   try {
     const postMessages = await PostMessage.find();
-    res.status(200).json(postMessages);
+    return res.status(200).json(postMessages);
   } catch (error) {
     res.status(404).json({ message: error.message });
   }
@@ -11,10 +11,15 @@ export const getPosts = async (req, res) => {
 
 export const createPost = async (req, res) => {
   try {
-    let post = req.body;
-    // console.log(post);
-    const newPost = await PostMessage.create(post);
-    res.status(201).json(newPost);
+    const { postData, name } = req.body;
+    if (!req.userId) return res.json({ message: "Unauthenticated" });
+    const newPost = await PostMessage.create({
+      ...postData,
+      name,
+      creator: req.userId,
+      createdAt: new Date().toISOString(),
+    });
+    return res.status(201).json(newPost);
   } catch (error) {
     res.status(409).json({ message: error.message });
   }
@@ -22,16 +27,20 @@ export const createPost = async (req, res) => {
 
 export const updatePost = async (req, res) => {
   const { id: _id } = req.params;
-  const post = req.body;
+  if (!req.userId) return res.json({ message: "Unauthenticated" });
+  const { postData, name } = req.body;
   if (!mongoose.Types.ObjectId.isValid(_id)) {
     return res.status(404).send("No Post with that id");
   }
   try {
-    const updatedPost = await PostMessage.findByIdAndUpdate(_id, post, {
-      new: true,
-    });
-    console.log(updatedPost);
-    res.json(updatedPost);
+    const updatedPost = await PostMessage.findByIdAndUpdate(
+      _id,
+      { ...postData, name },
+      {
+        new: true,
+      }
+    );
+    return res.json(updatedPost);
   } catch (error) {
     res.status(400).json({ message: error.message });
   }
@@ -39,12 +48,13 @@ export const updatePost = async (req, res) => {
 
 export const deletePost = async (req, res) => {
   const { id: _id } = req.params;
+  if (!req.userId) return res.json({ message: "Unauthenticated" });
   if (!mongoose.Types.ObjectId.isValid(_id)) {
     return res.status(404).send("No Post with that id");
   }
   try {
     await PostMessage.findByIdAndDelete(_id);
-    res.json({ message: "Post deleted successfully" });
+    return res.json({ message: "Post deleted successfully" });
   } catch (error) {
     res.status(400).json({ message: error.message });
   }
@@ -52,20 +62,26 @@ export const deletePost = async (req, res) => {
 
 export const likePost = async (req, res) => {
   const { id: _id } = req.params;
+  if (!req.userId) return res.json({ message: "Unauthenticated" });
   if (!mongoose.Types.ObjectId.isValid(_id)) {
     return res.status(404).send("No Post with that id");
   }
   try {
     const post = await PostMessage.findById(_id);
-    const updatedPost = await PostMessage.findByIdAndUpdate(
-      post._id,
-      {
-        likeCount: post.likeCount + 1,
-      },
-      { new: true }
-    );
-    res.json(updatedPost);
+    const userId = req.userId.toString();
+
+    const index = post.likes.findIndex((id) => id === userId);
+    if (index === -1) {
+      post.likes.push(req.userId);
+    } else {
+      post.likes = post.likes.filter((id) => id !== userId);
+    }
+
+    const updatedPost = await PostMessage.findByIdAndUpdate(post._id, post, {
+      new: true,
+    });
+    return res.status(200).json(updatedPost);
   } catch (error) {
-    res.status(400).jsno({ message: error.message });
+    res.status(400).json({ message: error.message });
   }
 };
